Initial commit of shiny application
Here's the initial commit of the shiny application part of the
assignment. The application attempts to fit entered data into a Gompertz
model, then predicts future values based on the model. It also allows
you to specify a date when you made a change that you believe will make
a difference to the growth patterns and calculates separate curves for
before and after the change.